+++ src/main/java/roomescape/booking/reservation/ReservationService.java
@@ -3,19 +3,9 @@ package roomescape.booking.reservation;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.auth.dto.LoginMember;
 import roomescape.booking.reservation.dto.AdminFilterReservationRequest;
-import roomescape.booking.reservation.dto.AdminReservationRequest;
-import roomescape.booking.reservation.dto.ReservationRequest;
 import roomescape.booking.reservation.dto.ReservationResponse;
-import roomescape.booking.schedule.Schedule;
-import roomescape.booking.schedule.ScheduleService;
-import roomescape.exception.custom.reason.reservation.ReservationConflictException;
 import roomescape.exception.custom.reason.reservation.ReservationNotFoundException;
-import roomescape.exception.custom.reason.reservation.ReservationPastDateException;
-import roomescape.exception.custom.reason.schedule.PastScheduleException;
-import roomescape.member.Member;
-import roomescape.member.MemberService;
 import roomescape.reservationtime.ReservationTime;
 import roomescape.theme.Theme;
 
@@ -27,28 +17,6 @@ import java.util.List;
 public class ReservationService {
 
     private final ReservationRepository reservationRepository;
-    private final ScheduleService scheduleService;
-    private final MemberService memberService;
-
-    @Transactional
-    public ReservationResponse create(final ReservationRequest request, final LoginMember loginMember) {
-        final Schedule schedule = scheduleService.findByDateAndTimeIdAndThemeId(request.date(), request.timeId(), request.themeId());
-        if (schedule.isPast()) {
-            throw new PastScheduleException();
-        }
-        final Member member = memberService.findByEmail(loginMember.email());
-        return getReservationResponse(schedule, member);
-    }
-
-    @Transactional
-    public ReservationResponse createForAdmin(final AdminReservationRequest request) {
-        final Schedule schedule = scheduleService.findByDateAndTimeIdAndThemeId(request.date(), request.timeId(), request.themeId());
-        if (schedule.isPast()) {
-            throw new PastScheduleException();
-        }
-        final Member member = memberService.findById(request.memberId());
-        return getReservationResponse(schedule, member);
-    }
 
     @Transactional(readOnly = true)
     public List<ReservationResponse> readAll() {
@@ -69,8 +37,7 @@ public class ReservationService {
 
     @Transactional(readOnly = true)
     public List<Reservation> findAllByEmail(final String email) {
-        Member member = memberService.findByEmail(email);
-        return reservationRepository.findAllByMember(member);
+        return reservationRepository.findAllByMember_Email(email);
     }
 
     @Transactional
@@ -102,25 +69,5 @@ public class ReservationService {
     @Transactional(readOnly = true)
     public boolean existsByReservationTime(final ReservationTime reservationTime) {
         return reservationRepository.existsBySchedule_ReservationTime(reservationTime);
-    }
-
-    private ReservationResponse getReservationResponse(Schedule schedule, final Member member) {
-        validateDuplicateReservation(schedule);
-        validatePast(schedule);
-        final Reservation notSavedReservation = new Reservation(member, schedule);
-        final Reservation savedReservation = reservationRepository.save(notSavedReservation);
-        return ReservationResponse.from(savedReservation);
-    }
-
-    private void validatePast(final Schedule schedule) {
-        if (schedule.isPast()) {
-            throw new ReservationPastDateException();
-        }
-    }
-
-    private void validateDuplicateReservation(final Schedule schedule) {
-        if (reservationRepository.existsBySchedule(schedule)) {
-            throw new ReservationConflictException();
-        }
     }
 }
